# server/analysis.openapi.yaml
openapi: 3.0.1
info:
  title: analysis
  description: ''
  version: 1.0.0
tags: []
paths:
  /predicts/groups:
    post:
      summary: group予測
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                talk_session_id:
                  type: string
              required:
                - user_id
                - talk_session_id
            example:
              user_id: '0'
              talk_session_id: '0'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
      security:
        - basic: []
  /reports/generates:
    post:
      summary: レポート作成
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                talk_session_id:
                  type: string
              required:
                - talk_session_id
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
      security:
        - basic: []
  /test:
    get:
      summary: テストAPI
      deprecated: false
      description: ''
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                    description: テキスト
                    default: hello
          headers: {}
      security: []
components:
  schemas: {}
  securitySchemes:
    basic:
      type: http
      scheme: basic
servers: []